pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "finance-agent-evaluator"
version = "2.0.0"
description = "Finance Agent Benchmark 2.0 - Green Agent (Evaluator)"
authors = [
    {name = "Finance Agent Benchmark Team"}
]
readme = "README.md"
requires-python = ">=3.11,<3.14"
dependencies = [
    "gymnasium>=1.0.0",
    "a2a-sdk[http-server]>=0.3.20",
    "pydantic>=2.0.0",
    "aiohttp>=3.9.0",
    "backoff>=2.2.0",
    "beautifulsoup4>=4.12.0",
    "python-dotenv>=1.0.0",
    "uvicorn>=0.27.0",
    "starlette>=0.35.0",
    "litellm>=1.0.0",
    "httpx>=0.28.0",
    "pandas>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]
phoenix = [
    "arize-phoenix-otel>=0.6.0",
    "openinference-instrumentation-litellm>=0.1.0",
    "openinference-semantic-conventions>=0.1.0",
]

[tool.uv]
package = true

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"
